pyproject.toml
[project]
name = "pms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aerich>=0.8.1",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "sqlmodel>=0.0.37",
    "tortoise>=0.1.1",
    "tortoise-orm>=1.1.6",
    "uvicorn>=0.41.0",
]
